Martijn added some things so the haptic driver + vibration motors work.
