+++ platformio.ini
@@ -14,5 +14,5 @@ board = lolin_d32_pro
 framework = arduino
 
 ; If Platformio gives a upload error, please specify this for your respective COM port
-upload_port = /dev/ttyUSB0
+upload_port = COM7
 upload_speed = 115200
